Mount is ALMOST done. Must create the tree inside the load function. For this, must decide how we are going to keep writing in the binary file.
My initial idea:

I) In partition.h, FILES_AND_SUBDIR is the position in the stream (binary file) of the first file or directory after the root of a given file system.
II) From FILES_AND_SUBDIR, we must save or retrieve the following file/directory infos:
  a) name (both) (Obs: we will save the absolute name in a char array with size 1024)
  b) size (file only)
  c) creation date (both) (For the dates, use the same idea of the functions implemented in mount.c. The size of the char array is 20)
  d) modification date (both)
  e) access date (both)
  f) data (file only)
III) To achieve (II), we must do a structure similar to this in partition.h.

NOTE: When constructing tree, must consider the case when the next file/directory may not be inside the parent directory.